wip:refactor: move core apis out of operator
