+++ apis/apis.go
@@ -1,0 +1,2 @@
+// Package apis contains Kubernetes API groups.
+package apis
