+++ operator/pkg/manifest/manifest_loader.go
@@ -18,7 +18,7 @@ import (
 	"context"
 	"fmt"
 
-	corev1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/apis/core/v1beta1"
+	corev1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/apis/core/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/k8s"
 
 	"k8s.io/apimachinery/pkg/runtime"
