+++ operator/pkg/manifest/repo_test.go
@@ -21,7 +21,7 @@ import (
 	"strings"
 	"testing"
 
-	corev1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/apis/core/v1beta1"
+	corev1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/apis/core/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/k8s"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/manifest"
 	testpaths "github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/test/util/paths"
