+++ operator/pkg/preflight/configconnectorcontextchecker.go
@@ -18,7 +18,7 @@ import (
 	"context"
 	"fmt"
 
-	corev1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/apis/core/v1beta1"
+	corev1beta1 "github.com/GoogleCloudPlatform/k8s-config-connector/apis/core/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/operator/pkg/k8s"
 
 	ctrl "sigs.k8s.io/controller-runtime"
